requirements.txt
# requests
kafka-python ==  2.0.2
opencv-contrib-python
Flask